Release version 3.4.0+1
Add explicit wildcard to match the .h files in Eigen/src/. The original
implementation relied on a bug in build2 which caused it to return files
with any extension for the `**.` pattern that is only supposed to return
files without any extension. This bug has been fixed for build2 0.16.0
(commit id 4ef4820f0f85).