Fix invalid include paths after installation
The relative include paths of the unsupported modules need to be kept
valid with respect to the relative location of Eigen header files.
Still, after installation, they should not pollute the system's
include folder with an 'unsupported' directory.
So, when installed, the Eigen headers and the unsupported Eigen modules
will be put in another directory 'Eigen' in the system's include folder
to provide the same relative locations as in the repository itself.

Issues: #2